@@ -28,35 +28,32 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  cached_network_image: ^3.4.1
+  connectivity_plus: ^7.0.0
+  cupertino_icons: ^1.0.8
+  dartz: ^0.10.1
+  dio: ^5.8.0+1
+  equatable: ^2.0.7
   flutter:
     sdk: flutter
-  flutter_screenutil: ^5.9.3
+  flutter_bloc: ^9.1.1
   flutter_localizations:
     sdk: flutter
-  flutter_bloc: ^9.1.1
-  shared_preferences: ^2.5.3
-  dio: ^5.8.0+1
-  
+  flutter_screenutil: ^5.9.3
+  flutter_translate: ^3.1.0
+  font_awesome_flutter: ^10.12.0
   get_it: ^8.0.3
   go_router: ^16.0.0
-  flutter_translate: ^3.1.0
-  dartz: ^0.10.1
-  cached_network_image: ^3.4.1
-  top_snackbar_flutter: ^3.3.0
-  connectivity_plus: ^7.0.0
   internet_connection_checker: ^1.0.0+1
   logger: ^2.6.2
-  equatable: ^2.0.7
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
+  shared_preferences: ^2.5.3
+  top_snackbar_flutter: ^3.3.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   flutter_launcher_icons: ^0.14.4
   flutter_lints: ^5.0.0
+  flutter_test:
+    sdk: flutter
 
 flutter_icons:
   android: true
@@ -81,7 +78,6 @@ flutter_icons:
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter packages.
 flutter:
 
@@ -97,10 +93,8 @@ flutter:
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/to/asset-from-package
-
   # Alexandria Font Family
   fonts:
     - family: Alexandria
